requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
console-menu==0.6.0
cycler==0.10.0
elasticsearch==7.15.0
fake-useragent==0.1.11
future==0.18.2
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.1
multidict==5.2.0
mysql-connector-python==8.0.23
numpy==1.20.1
oauthlib==3.1.1
pandas==1.2.3
pefile==2021.5.24
Pillow==8.4.0
protobuf==3.19.0
pycares==4.0.0
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-socks==1.2.4
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
tweepy==4.2.0
twint==2.1.20
TwitterAPI==2.7.5
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
